lsp: fix error-recovering parser hangs; lexer: fix int-literal overflow UB
Fuzzing the front-end surfaced three real bugs, fixed here:

- LSP parser infinite-loops on a `when` statement whose discriminant is a
  member-access expression rooted at `self` (e.g. `when self.kind`), as in
  the valid examples/lox/value.roxy. parse_when_stmt hard-coded an
  Identifier(.Identifier)* discriminant and could not consume the `self`
  keyword, leaving the cases loop parked. Now parses the discriminant as a
  struct-literal-suppressed expression, matching the compiler parser.

- LSP parser infinite-loops on a stray leading token (`}`, `"`, `,`, `::`,
  `0x`, ...): parse_primary returns an error node without consuming, and the
  unbounded declaration loops had no forward-progress guarantee. Added
  progress backstops to parse_program / parse_block_stmt and the two `when`
  recovery loops (advance one token whenever an iteration consumed nothing).

- Lexer signed-overflow UB on out-of-range integer literals
  (e.g. 99999999999999999999): accumulate in u64 (defined modular wraparound)
  and store the two's-complement bits into the i64 token value.

All 1367 non-C tests pass (302k assertions); 123 LSP/CST/indexer cases
unchanged.

+++ src/roxy/shared/lexer.cpp
@@ -307,7 +307,13 @@ Token Lexer::scan_number() {
     token.start = m_source + m_start;
 
     bool is_float = false;
-    i64 int_value = 0;
+    // Accumulate the integer value as unsigned: unsigned overflow is defined
+    // (modular wraparound), whereas signed overflow is UB. A literal larger
+    // than i64 max (e.g. 99999999999999999999) therefore wraps deterministically
+    // instead of triggering undefined behavior; the resulting bit pattern is
+    // stored back into the i64 token value (well-defined two's-complement
+    // conversion in C++20). Range diagnostics, if wanted, belong in sema.
+    u64 int_value = 0;
     f64 float_value = 0.0;
 
     // Check for hex, binary, or octal prefix
@@ -395,7 +401,7 @@ suffixes:
         token.float_value = float_value;
     } else {
         token.kind = TokenKind::IntLiteral;
-        token.int_value = int_value;
+        token.int_value = (i64)int_value;
     }
 
     return token;
